Move patch helper files
