+++ SGGL-Diablo-Knowledge-Library/src/patch_helper/buffer_patch.c
