+++ SGGL-Diablo-Knowledge-Library/src/patch_helper/pe_header.c
